// apps/sovoli.com/src/modules/data/organisations/private-schools/guyana/metadata.generated.ts
export interface PrivateSchoolMetadata {
  name: string;
  address?: string;
  website?: string;
  phones: string[];
  placeId?: string;
  facebook?: {
    url: string;
    lastUpdated?: string;
  };
  otherSources?: string[];
}

export const PRIVATE_SCHOOL_GUYANA_METADATA: Record<string, PrivateSchoolMetadata> = {
  "academyexcellence": {
    "name": "Academy of Excellence",
    "phones": [],
    "address": "VP6J+22V, Cornelia Ida, Guyana",
    "placeId": "ChIJlSUlKWXnr40R7Q6ZUCJHwsM",
    "facebook": {
      "url": "https://www.facebook.com/Academy-of-Excellence-833797406654539/",
      "lastUpdated": "September 3, 2014"
    }
  },
// ...
} as const;